Revert "fix(server): 将 points 计数改为 long"
This reverts commit 2a723e8a79d5bed19cf5f7d8f066561079205493.

+++ server/src/main/kotlin/infra/web/WebNovel.kt
@@ -91,7 +91,7 @@ class WebNovel(
     val type: WebNovelType = WebNovelType.连载中,
     val attentions: List<WebNovelAttention> = emptyList(),
     val keywords: List<String> = emptyList(),
-    val points: Long?,
+    val points: Int?,
     val totalCharacters: Int? = null,
     val introductionJp: String,
     val introductionZh: String? = null,
